% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{helper}
\alias{helper}
\title{Insert formatted headers}
\arguments{
\item{lw}{Integer. Width at which to split the lines. Default = 70.}

\item{name}{Character. Default name to use as author. Set to NULL
if no default, else will choose from option helper.name.}
}
\description{
`helper()` is a function that runs a Shiny gadget in the viewpane
to allow for the formatting and insertion of headers.
}
