% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inherit_beamer_presentation}
\alias{inherit_beamer_presentation}
\title{#' @import redoc
# Call rmarkdown::word_document and mark the return value as inheriting word_document
inherit_redoc_document <- function(...) {
  fmt <- redoc::redoc(..., wrappers = list(latexwrap, citationwrap, rawblockwrap, rawspanwrap))
  fmt$inherits <- "redoc_document"
  fmt
}}
\usage{
inherit_beamer_presentation(...)
}
\description{
#' @import redoc
# Call rmarkdown::word_document and mark the return value as inheriting word_document
inherit_redoc_document <- function(...) {
  fmt <- redoc::redoc(..., wrappers = list(latexwrap, citationwrap, rawblockwrap, rawspanwrap))
  fmt$inherits <- "redoc_document"
  fmt
}
}
