% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_chunk.R
\name{make_chunk}
\alias{make_chunk}
\title{Insert formatted headers}
\usage{
make_chunk()
}
\description{
`make_chunk()` is a function that allows a user to easily make highlighted
text into an rmarkdown code chunk
}
