% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_ggplot.R
\name{my_ggplot}
\alias{my_ggplot}
\alias{my_scale_color}
\alias{my_scale_fill}
\alias{my_theme}
\title{Customize ggplot}
\usage{
my_scale_color(
  type = "div",
  data = "cont",
  other = FALSE,
  bw = FALSE,
  direction = 1,
  ...
)

my_scale_fill(
  type = "div",
  data = "cont",
  other = FALSE,
  bw = FALSE,
  direction = 1,
  ...
)

my_theme(..., font = TRUE, axis = "c")
}
\arguments{
\item{type}{One of "div", "seq", or "qual": the type of palette desired.}

\item{data}{One of "cont" or "discrete: the type of data being plotted.}

\item{other}{Logical. Use the "other" version of the palette if true.}

\item{bw}{Logical. Print a black and white version of the figure.}

\item{font}{Logical. Whether the figure should be plotted with the Roboto Condensed font (requires installation).}
}
\description{
These functions provide customization for ggplot2.
}
