% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{my_tables}
\alias{my_tables}
\alias{my_df_html}
\alias{my_gt_table1}
\title{Tables for printing}
\usage{
my_df_html(df)

my_gt_table1(.data, strata, allVars, factorVars, test = FALSE,
  printToggle = FALSE, noSpaces = TRUE, showAllLevels = TRUE, ...)
}
\arguments{
\item{.data}{a dataframe}

\item{strata}{a character vector of variables to stratify by (if NULL, only overall table)}

\item{allVars}{a character vector of variables to be rows in the table}

\item{factorVars}{a character vector of factor variables}
}
\description{
This function makes it easy to print all output from an rmarkdown document in a certain way.

other parameters to print.TableOne can be passed as well
}
