% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_project.R
\name{new_project}
\alias{new_project}
\title{Create a new project}
\usage{
new_project(
  path,
  use_targets = TRUE,
  use_renv = TRUE,
  use_git = TRUE,
  packages = c("tidyverse"),
  ...
)
}
\description{
Create a new project
}
