% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline_targets.R
\name{outline_targets}
\alias{outline_targets}
\title{Create a targets outline}
\usage{
outline_targets()
}
\description{
`outline_targets()` is used to clean up `_targets.R` files (in the current
working directory) by inserting headers at every target and set of targets
so that the RStudio outline function can be used.
}
